Follow system fully functional

+++ public/javascripts/short-delete.js
@@ -1,7 +1,7 @@
 window.addEventListener('DOMContentLoaded', () => {
     const deleteButtons = Array.from(document.querySelectorAll('.delete'));
-    deleteButtons.forEach( button => {
-        button.addEventListener('click',() => {
+    deleteButtons.forEach(button => {
+        button.addEventListener('click', () => {
             fetch(`/shorts/${button.id}`, {
                 method: 'DELETE',
             });
@@ -11,14 +11,15 @@ window.addEventListener('DOMContentLoaded', () => {
 
     const pageDeleteButton = document.querySelector('.delete-page');
 
-    pageDeleteButton.addEventListener('click', event => {
-        console.log(window.location.href);
-        fetch(`/shorts/${event.target.id}`, {
-            method: 'DELETE',
-        }).then( res => {
-            if(res.redirected) {
-                window.location.href= '/shorts'
-            }
+    if (pageDeleteButton) {
+        pageDeleteButton.addEventListener('click', event => {
+            fetch(`/shorts/${event.target.id}`, {
+                method: 'DELETE',
+            }).then(res => {
+                if (res.redirected) {
+                    window.location.href = '/shorts'
+                }
+            });
         });
-    });
+    }
 });
